@@ -1,8 +1,9 @@
 use std::fs;
 
-use zed::lsp::{Completion, CompletionKind, Symbol, SymbolKind};
-use zed::{CodeLabel, CodeLabelSpan, LanguageServerId};
-use zed_extension_api::{self as zed, Result};
+use zed_extension_api::{
+    self as zed, CodeLabel, CodeLabelSpan, LanguageServerId, Result, Worktree,
+    lsp::{Completion, CompletionKind, Symbol, SymbolKind},
+};
 
 use crate::language_servers::util;
 
@@ -22,16 +23,16 @@ impl NextLs {
     pub fn language_server_binary_path(
         &mut self,
         language_server_id: &LanguageServerId,
-        worktree: &zed::Worktree,
+        worktree: &Worktree,
     ) -> Result<String> {
-        if let Some(path) = worktree.which("nextls") {
-            return Ok(path);
+        if let Some(binary_path) = worktree.which(&Self::LANGUAGE_SERVER_ID.replace("-", "")) {
+            return Ok(binary_path);
         }
 
-        if let Some(path) = &self.cached_binary_path {
-            if fs::metadata(path).map_or(false, |stat| stat.is_file()) {
-                return Ok(path.clone());
-            }
+        if let Some(binary_path) = &self.cached_binary_path
+            && fs::metadata(binary_path).is_ok_and(|stat| stat.is_file())
+        {
+            return Ok(binary_path.clone());
         }
 
         zed::set_language_server_installation_status(
@@ -77,7 +78,7 @@ impl NextLs {
 
         let binary_path = format!("{version_dir}/next-ls");
 
-        if !fs::metadata(&binary_path).map_or(false, |stat| stat.is_file()) {
+        if !fs::metadata(&binary_path).is_ok_and(|stat| stat.is_file()) {
             zed::set_language_server_installation_status(
                 language_server_id,
                 &zed::LanguageServerInstallationStatus::Downloading,
